dash: For first segments, don't choose the lowest bitrate.
For some streams, the lowest bitrate is an audio stream only. We will
fall back to a more appropriate stream after a few chunks.

+++ modules/stream_filter/dash/adaptationlogic/AbstractAdaptationLogic.h
@@ -52,7 +52,7 @@ namespace dash
                 long                        getBpsLastChunk         ();
 
             private:
-                long                    bpsAvg;
+                int                     bpsAvg;
                 long                    bpsLastChunk;
                 dash::mpd::IMPDManager  *mpdManager;
         };
